api_v2/auth/requirements.txt
Flask
Flask-RESTful
pytest
pytest-cov
coveralls
python-keycloak
prometheus-client